@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFramework>net7.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<ApplicationIcon>Images\BaoXia_600x600@1x.ico</ApplicationIcon>
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
 		<Company>石家庄宝匣软件科技有限公司</Company>
 		<Description>宝匣软件 .Net Core 公用工具集。</Description>
 		<PackageLicenseExpression>MIT</PackageLicenseExpression>
-		<Version>7.0.19</Version>
+		<Version>8.0.0</Version>
 		<PackageIcon>BaoXia_600x600@1x.png</PackageIcon>
 		<PackOnBuild>true</PackOnBuild>
-		<FileVersion>7.0.19</FileVersion>
+		<FileVersion>8.0.0</FileVersion>
 		<ReleaseVersion>1.1.11</ReleaseVersion>
-		<AssemblyVersion>7.0.19</AssemblyVersion>
+		<AssemblyVersion>8.0.0</AssemblyVersion>
 		<Nullable>enable</Nullable>
 		<Authors>石家庄宝匣软件科技有限公司</Authors>
 		<Title>宝匣软件 - 基础工具包</Title>
