@@ -1,1078 +1,1290 @@
 ﻿using BaoXia.Utils.Constants;
 using System;
 
-namespace BaoXia.Utils.Extensions
+namespace BaoXia.Utils.Extensions;
+
+/// <summary>
+/// “DateTime”扩展类。
+/// </summary>
+public static class DateTimeExtension
 {
 	/// <summary>
-	/// “DateTime”扩展类。
+	/// 通过指定距离1970年1月1日零时的毫秒数，创建当前时区的时间。
 	/// </summary>
-	public static class DateTimeExtension
+	/// <param name="millisecondsFrom1970">距离1970年1月1日零时的毫秒数。</param>
+	/// <returns>距离1970年1月1日零时的毫秒数对应的当前时区时间。</returns>
+	public static DateTime DateTimeWithMillisecondsFrom1970(
+	    long millisecondsFrom1970)
 	{
-		/// <summary>
-		/// 通过指定距离1970年1月1日零时的毫秒数，创建当前时区的时间。
-		/// </summary>
-		/// <param name="millisecondsFrom1970">距离1970年1月1日零时的毫秒数。</param>
-		/// <returns>距离1970年1月1日零时的毫秒数对应的当前时区时间。</returns>
-		public static DateTime DateTimeWithMillisecondsFrom1970(
-		    long millisecondsFrom1970)
+		var dateTime = new DateTime(1970, 1, 1);
 		{
-			var dateTime = new DateTime(1970, 1, 1);
+			dateTime = dateTime.AddMilliseconds(millisecondsFrom1970);
+			//
+			dateTime = dateTime.ToLocalTime();
+		}
+		return dateTime;
+	}
+
+	/// <summary>
+	/// 获取当前时间的UTC时间戳，当前时间的UTC时间距离1970年1月1日零时的毫秒数。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>返回当前时间的UTC时间距离1970年1月1日零时的毫秒数。</returns>
+	public static long MillisecondsFrom1970(this DateTime dateTime)
+	{
+		var dateTimeUtc = TimeZoneInfo.ConvertTimeToUtc(dateTime);
+		var zeroTime = new DateTime(1970, 1, 1);
+
+		var millisecondsFrom1970 = (long)(dateTimeUtc - zeroTime).TotalMilliseconds;
+		{ }
+		return millisecondsFrom1970;
+	}
+
+	/// <summary>
+	/// 返回当前时间零点的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间零点的时间对象。</returns>
+	public static DateTime ZeroOfThisDay(this DateTime dateTime)
+	{
+		return new DateTime(dateTime.Year, dateTime.Month, dateTime.Day);
+	}
+
+	/// <summary>
+	/// 返回当前时间前一天零点的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间零点的时间对象。</returns>
+	public static DateTime ZeroOfPrevDay(this DateTime dateTime)
+	{
+		var prevDay = dateTime.AddDays(-1);
+		{ }
+		return new DateTime(prevDay.Year, prevDay.Month, prevDay.Day);
+	}
+
+	/// <summary>
+	/// 返回当前时间后一天零点的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间零点的时间对象。</returns>
+	public static DateTime ZeroOfNextDay(this DateTime dateTime)
+	{
+		var nextDay = dateTime.AddDays(1);
+		{ }
+		return new DateTime(nextDay.Year, nextDay.Month, nextDay.Day);
+	}
+
+	/// <summary>
+	/// 返回当前时间所属周（以周日为第一天的周）的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间所属周（以周日为第一天的周）的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfThisWeekStartsWithSunday(this DateTime dateTime)
+	{
+		var dayOfWeek = dateTime.DayOfWeek;
+		var firstDayOfThisWeek = dateTime.AddDays(0 - (int)dayOfWeek);
+		{
+			firstDayOfThisWeek = new DateTime(
+				firstDayOfThisWeek.Year,
+				firstDayOfThisWeek.Month,
+				firstDayOfThisWeek.Day);
+		}
+		return firstDayOfThisWeek;
+	}
+
+	/// <summary>
+	/// 返回当前时间上一周（以周日为第一天的周）的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间上一周（以周日为第一天的周）的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfPrevWeekStartsWithSunday(this DateTime dateTime)
+	{
+		var prevWeek = dateTime.AddDays(-TimeConstants.DaysToWeek);
+		var firstDayOfPrevWeek = prevWeek.FirstDayOfThisWeekStartsWithSunday();
+		{ }
+		return firstDayOfPrevWeek;
+	}
+
+	/// <summary>
+	/// 返回当前时间下一周（以周日为第一天的周）的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间下一周（以周日为第一天的周）的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfNextWeekStartsWithSunday(this DateTime dateTime)
+	{
+
+		var nextWeek = dateTime.AddDays(TimeConstants.DaysToWeek);
+		var firstDayOfNextWeek = nextWeek.FirstDayOfThisWeekStartsWithSunday();
+		{ }
+		return firstDayOfNextWeek;
+	}
+
+	/// <summary>
+	/// 返回当前时间所属周（以周一为第一天的周）的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间所属周（以周一为第一天的周）的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfThisWeekStartsWithMonday(this DateTime dateTime)
+	{
+		var dayOfWeek = dateTime.DayOfWeek;
+		var firstDayOfThisWeek = dateTime.AddDays(0 - (int)dayOfWeek + 1);
+		{
+			firstDayOfThisWeek = new DateTime(
+				firstDayOfThisWeek.Year,
+				firstDayOfThisWeek.Month,
+				firstDayOfThisWeek.Day);
+		}
+		return firstDayOfThisWeek;
+	}
+
+	/// <summary>
+	/// 返回当前时间上一周（以周一为第一天的周）的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间上一周（以周一为第一天的周）的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfPrevWeekStartsWithMonday(this DateTime dateTime)
+	{
+		var prevWeek = dateTime.AddDays(-TimeConstants.DaysToWeek);
+		var firstDayOfPrevWeek = prevWeek.FirstDayOfThisWeekStartsWithMonday();
+		{ }
+		return firstDayOfPrevWeek;
+	}
+
+	/// <summary>
+	/// 返回当前时间下一周（以周一为第一天的周）的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间下一周（以周一为第一天的周）的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfNextWeekStartsWithMonday(this DateTime dateTime)
+	{
+		var nextWeek = dateTime.AddDays(TimeConstants.DaysToWeek);
+		var firstDayOfNextWeek = nextWeek.FirstDayOfThisWeekStartsWithMonday();
+		{ }
+		return firstDayOfNextWeek;
+	}
+
+	/// <summary>
+	/// 返回当前时间所属月份的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间所属月份的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfThisMonth(this DateTime dateTime)
+	{
+		return new DateTime(dateTime.Year, dateTime.Month, 1);
+	}
+
+	/// <summary>
+	/// 返回当前时间上一个月份的第一天的时间（零点）对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间上一个月份的第一天的时间（零点）对象。</returns>
+	public static DateTime FirstDayOfPrevMonth(this DateTime dateTime)
+	{
+		return dateTime.AddMonths(-1).FirstDayOfThisMonth();
+	}
+
+	/// <summary>
+	/// 返回当前时间下一个月份的第一天的时间（零点）对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间下一个月份的第一天的时间（零点）对象。</returns>
+	public static DateTime FirstDayOfNextMonth(this DateTime dateTime)
+	{
+		return dateTime.AddMonths(1).FirstDayOfThisMonth();
+	}
+
+	/// <summary>
+	/// 返回当前时间所属季度的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间所属季度的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfThisSession(this DateTime dateTime)
+	{
+		var thisMonth = dateTime.Month;
+		var firstMonthOfThisSession = thisMonth;
+		if (thisMonth >= TimeConstants.FirstMonthOfSession1 && thisMonth <= TimeConstants.LastMonthOfSession1)
+		{
+			firstMonthOfThisSession = 1;
+		}
+		else if (thisMonth >= TimeConstants.FirstMonthOfSession2 && thisMonth <= TimeConstants.LastMonthOfSession2)
+		{
+			firstMonthOfThisSession = 4;
+		}
+		else if (thisMonth >= TimeConstants.FirstMonthOfSession3 && thisMonth <= TimeConstants.LastMonthOfSession3)
+		{
+			firstMonthOfThisSession = 7;
+		}
+		else if (thisMonth >= TimeConstants.FirstMonthOfSession4 && thisMonth <= TimeConstants.LastMonthOfSession4)
+		{
+			firstMonthOfThisSession = 10;
+		}
+		var firstDayOfThisSession = new DateTime(dateTime.Year, firstMonthOfThisSession, 1);
+		{ }
+		return firstDayOfThisSession;
+	}
+
+	/// <summary>
+	/// 返回当前时间上一个季度的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间上一个季度的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfPrevSession(this DateTime dateTime)
+	{
+		return dateTime
+			.AddMonths(-TimeConstants.MonthsToSession)
+			.FirstDayOfThisSession();
+	}
+
+	/// <summary>
+	/// 返回当前时间下一个季度的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间下一个季度的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfNextSession(this DateTime dateTime)
+	{
+		return dateTime
+			.AddMonths(+TimeConstants.MonthsToSession)
+			.FirstDayOfThisSession();
+	}
+
+
+	/// <summary>
+	/// 返回当前时间所属年份的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间所属年份的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfThisYear(this DateTime dateTime)
+	{
+		return new DateTime(dateTime.Year, 1, 1);
+	}
+
+	/// <summary>
+	/// 返回当前时间上一年的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间上一年的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfPrevYear(this DateTime dateTime)
+	{
+		return new DateTime(dateTime.Year - 1, 1, 1);
+	}
+
+	/// <summary>
+	/// 返回当前时间下一年的第一天（零点）的时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <returns>当前时间下一年的第一天（零点）的时间对象。</returns>
+	public static DateTime FirstDayOfNextYear(this DateTime dateTime)
+	{
+		return new DateTime(dateTime.Year + 1, 1, 1);
+	}
+
+	/// <summary>
+	/// 在指定的时间精度上比较两个时间对象。
+	/// </summary>
+	/// <param name="dateTime">当前时间对象。</param>
+	/// <param name="anotherDateTime">另一个时间对象。</param>
+	/// <param name="compareFieldMin">指定的时间精度，类型为：DateTimeField。</param>
+	/// <returns>“dateTime”小于“anotherDateTime”时，返回“-1”；“dateTime”等于“anotherDateTime”时，返回“0”；“dateTime”大于“anotherDateTime”时，返回“1”。</returns>
+	public static int CompareTo(
+		this DateTime dateTime,
+		DateTime anotherDateTime,
+		DateTimeField compareFieldMin = DateTimeField.Millisecond)
+	{
+		if (compareFieldMin >= DateTimeField.Year)
+		{
+			if (dateTime.Year > anotherDateTime.Year)
 			{
-				dateTime = dateTime.AddMilliseconds(millisecondsFrom1970);
-				//
-				dateTime = dateTime.ToLocalTime();
+				return 1;
 			}
-			return dateTime;
-		}
-
-		/// <summary>
-		/// 获取当前时间的UTC时间戳，当前时间的UTC时间距离1970年1月1日零时的毫秒数。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>返回当前时间的UTC时间距离1970年1月1日零时的毫秒数。</returns>
-		public static long MillisecondsFrom1970(this DateTime dateTime)
-		{
-			var dateTimeUtc = TimeZoneInfo.ConvertTimeToUtc(dateTime);
-			var zeroTime = new DateTime(1970, 1, 1);
-
-			var millisecondsFrom1970 = (long)(dateTimeUtc - zeroTime).TotalMilliseconds;
-			{ }
-			return millisecondsFrom1970;
-		}
-
-		/// <summary>
-		/// 返回当前时间零点的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间零点的时间对象。</returns>
-		public static DateTime ZeroOfThisDay(this DateTime dateTime)
-		{
-			return new DateTime(dateTime.Year, dateTime.Month, dateTime.Day);
-		}
-
-		/// <summary>
-		/// 返回当前时间前一天零点的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间零点的时间对象。</returns>
-		public static DateTime ZeroOfPrevDay(this DateTime dateTime)
-		{
-			var prevDay = dateTime.AddDays(-1);
-			{ }
-			return new DateTime(prevDay.Year, prevDay.Month, prevDay.Day);
-		}
-
-		/// <summary>
-		/// 返回当前时间后一天零点的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间零点的时间对象。</returns>
-		public static DateTime ZeroOfNextDay(this DateTime dateTime)
-		{
-			var nextDay = dateTime.AddDays(1);
-			{ }
-			return new DateTime(nextDay.Year, nextDay.Month, nextDay.Day);
-		}
-
-		/// <summary>
-		/// 返回当前时间所属周（以周日为第一天的周）的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间所属周（以周日为第一天的周）的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfThisWeekStartsWithSunday(this DateTime dateTime)
-		{
-			var dayOfWeek = dateTime.DayOfWeek;
-			var firstDayOfThisWeek = dateTime.AddDays(0 - (int)dayOfWeek);
+			else if (dateTime.Year < anotherDateTime.Year)
 			{
-				firstDayOfThisWeek = new DateTime(
-					firstDayOfThisWeek.Year,
-					firstDayOfThisWeek.Month,
-					firstDayOfThisWeek.Day);
+				return -1;
 			}
-			return firstDayOfThisWeek;
 		}
-
-		/// <summary>
-		/// 返回当前时间上一周（以周日为第一天的周）的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间上一周（以周日为第一天的周）的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfPrevWeekStartsWithSunday(this DateTime dateTime)
+		if (compareFieldMin >= DateTimeField.Month)
 		{
-			var prevWeek = dateTime.AddDays(-TimeConstants.DaysToWeek);
-			var firstDayOfPrevWeek = prevWeek.FirstDayOfThisWeekStartsWithSunday();
-			{ }
-			return firstDayOfPrevWeek;
-		}
-
-		/// <summary>
-		/// 返回当前时间下一周（以周日为第一天的周）的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间下一周（以周日为第一天的周）的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfNextWeekStartsWithSunday(this DateTime dateTime)
-		{
-
-			var nextWeek = dateTime.AddDays(TimeConstants.DaysToWeek);
-			var firstDayOfNextWeek = nextWeek.FirstDayOfThisWeekStartsWithSunday();
-			{ }
-			return firstDayOfNextWeek;
-		}
-
-		/// <summary>
-		/// 返回当前时间所属周（以周一为第一天的周）的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间所属周（以周一为第一天的周）的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfThisWeekStartsWithMonday(this DateTime dateTime)
-		{
-			var dayOfWeek = dateTime.DayOfWeek;
-			var firstDayOfThisWeek = dateTime.AddDays(0 - (int)dayOfWeek + 1);
+			if (dateTime.Month > anotherDateTime.Month)
 			{
-				firstDayOfThisWeek = new DateTime(
-					firstDayOfThisWeek.Year,
-					firstDayOfThisWeek.Month,
-					firstDayOfThisWeek.Day);
+				return 1;
 			}
-			return firstDayOfThisWeek;
-		}
-
-		/// <summary>
-		/// 返回当前时间上一周（以周一为第一天的周）的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间上一周（以周一为第一天的周）的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfPrevWeekStartsWithMonday(this DateTime dateTime)
-		{
-			var prevWeek = dateTime.AddDays(-TimeConstants.DaysToWeek);
-			var firstDayOfPrevWeek = prevWeek.FirstDayOfThisWeekStartsWithMonday();
-			{ }
-			return firstDayOfPrevWeek;
-		}
-
-		/// <summary>
-		/// 返回当前时间下一周（以周一为第一天的周）的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间下一周（以周一为第一天的周）的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfNextWeekStartsWithMonday(this DateTime dateTime)
-		{
-			var nextWeek = dateTime.AddDays(TimeConstants.DaysToWeek);
-			var firstDayOfNextWeek = nextWeek.FirstDayOfThisWeekStartsWithMonday();
-			{ }
-			return firstDayOfNextWeek;
-		}
-
-		/// <summary>
-		/// 返回当前时间所属月份的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间所属月份的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfThisMonth(this DateTime dateTime)
-		{
-			return new DateTime(dateTime.Year, dateTime.Month, 1);
-		}
-
-		/// <summary>
-		/// 返回当前时间上一个月份的第一天的时间（零点）对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间上一个月份的第一天的时间（零点）对象。</returns>
-		public static DateTime FirstDayOfPrevMonth(this DateTime dateTime)
-		{
-			return dateTime.AddMonths(-1).FirstDayOfThisMonth();
-		}
-
-		/// <summary>
-		/// 返回当前时间下一个月份的第一天的时间（零点）对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间下一个月份的第一天的时间（零点）对象。</returns>
-		public static DateTime FirstDayOfNextMonth(this DateTime dateTime)
-		{
-			return dateTime.AddMonths(1).FirstDayOfThisMonth();
-		}
-
-		/// <summary>
-		/// 返回当前时间所属季度的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间所属季度的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfThisSession(this DateTime dateTime)
-		{
-			var thisMonth = dateTime.Month;
-			var firstMonthOfThisSession = thisMonth;
-			if (thisMonth >= TimeConstants.FirstMonthOfSession1 && thisMonth <= TimeConstants.LastMonthOfSession1)
+			else if (dateTime.Month < anotherDateTime.Month)
 			{
-				firstMonthOfThisSession = 1;
+				return -1;
 			}
-			else if (thisMonth >= TimeConstants.FirstMonthOfSession2 && thisMonth <= TimeConstants.LastMonthOfSession2)
-			{
-				firstMonthOfThisSession = 4;
-			}
-			else if (thisMonth >= TimeConstants.FirstMonthOfSession3 && thisMonth <= TimeConstants.LastMonthOfSession3)
-			{
-				firstMonthOfThisSession = 7;
-			}
-			else if (thisMonth >= TimeConstants.FirstMonthOfSession4 && thisMonth <= TimeConstants.LastMonthOfSession4)
-			{
-				firstMonthOfThisSession = 10;
-			}
-			var firstDayOfThisSession = new DateTime(dateTime.Year, firstMonthOfThisSession, 1);
-			{ }
-			return firstDayOfThisSession;
 		}
-
-		/// <summary>
-		/// 返回当前时间上一个季度的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间上一个季度的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfPrevSession(this DateTime dateTime)
+		if (compareFieldMin >= DateTimeField.Day)
 		{
-			return dateTime
-				.AddMonths(-TimeConstants.MonthsToSession)
-				.FirstDayOfThisSession();
+			if (dateTime.Day > anotherDateTime.Day)
+			{
+				return 1;
+			}
+			else if (dateTime.Day < anotherDateTime.Day)
+			{
+				return -1;
+			}
 		}
-
-		/// <summary>
-		/// 返回当前时间下一个季度的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间下一个季度的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfNextSession(this DateTime dateTime)
+		if (compareFieldMin >= DateTimeField.Hour)
 		{
-			return dateTime
-				.AddMonths(+TimeConstants.MonthsToSession)
-				.FirstDayOfThisSession();
+			if (dateTime.Hour > anotherDateTime.Hour)
+			{
+				return 1;
+			}
+			else if (dateTime.Hour < anotherDateTime.Hour)
+			{
+				return -1;
+			}
 		}
-
-
-		/// <summary>
-		/// 返回当前时间所属年份的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间所属年份的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfThisYear(this DateTime dateTime)
+		if (compareFieldMin >= DateTimeField.Minute)
 		{
-			return new DateTime(dateTime.Year, 1, 1);
+			if (dateTime.Minute > anotherDateTime.Minute)
+			{
+				return 1;
+			}
+			else if (dateTime.Minute < anotherDateTime.Minute)
+			{
+				return -1;
+			}
 		}
-
-		/// <summary>
-		/// 返回当前时间上一年的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间上一年的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfPrevYear(this DateTime dateTime)
+		if (compareFieldMin >= DateTimeField.Second)
 		{
-			return new DateTime(dateTime.Year - 1, 1, 1);
+			if (dateTime.Second > anotherDateTime.Second)
+			{
+				return 1;
+			}
+			else if (dateTime.Second < anotherDateTime.Second)
+			{
+				return -1;
+			}
 		}
-
-		/// <summary>
-		/// 返回当前时间下一年的第一天（零点）的时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <returns>当前时间下一年的第一天（零点）的时间对象。</returns>
-		public static DateTime FirstDayOfNextYear(this DateTime dateTime)
+		if (compareFieldMin >= DateTimeField.Millisecond)
 		{
-			return new DateTime(dateTime.Year + 1, 1, 1);
+			if (dateTime.Millisecond > anotherDateTime.Millisecond)
+			{
+				return 1;
+			}
+			else if (dateTime.Millisecond < anotherDateTime.Millisecond)
+			{
+				return -1;
+			}
 		}
+		return 0;
+	}
 
-		/// <summary>
-		/// 在指定的时间精度上比较两个时间对象。
-		/// </summary>
-		/// <param name="dateTime">当前时间对象。</param>
-		/// <param name="anotherDateTime">另一个时间对象。</param>
-		/// <param name="compareFieldMin">指定的时间精度，类型为：DateTimeField。</param>
-		/// <returns>“dateTime”小于“anotherDateTime”时，返回“-1”；“dateTime”等于“anotherDateTime”时，返回“0”；“dateTime”大于“anotherDateTime”时，返回“1”。</returns>
-		public static int CompareTo(
-			this DateTime dateTime,
-			DateTime anotherDateTime,
-			DateTimeField compareFieldMin = DateTimeField.Millisecond)
+
+	/// <summary>
+	/// 在指定的时间周期内比较两个时间对象，如：指定的时间范围为“Year”时，则比较时忽略时间对象的年份值。
+	/// </summary>
+	/// <param name="dateTime"></param>
+	/// <param name="anotherDateTime"></param>
+	/// <param name="compareCycle"></param>
+	/// <returns></returns>
+	public static int CompareTo(
+		this DateTime dateTime,
+		DateTime anotherDateTime,
+		DateTimeCycle compareCycle)
+	{
+		switch (compareCycle)
 		{
-			if (compareFieldMin >= DateTimeField.Year)
-			{
-				if (dateTime.Year > anotherDateTime.Year)
+			default:
+			case DateTimeCycle.All:
+			case DateTimeCycle.Century:
 				{
-					return 1;
+					return dateTime.CompareTo(anotherDateTime);
 				}
-				else if (dateTime.Year < anotherDateTime.Year)
+			case DateTimeCycle.Year:
 				{
-					return -1;
-				}
-			}
-			if (compareFieldMin >= DateTimeField.Month)
-			{
-				if (dateTime.Month > anotherDateTime.Month)
-				{
-					return 1;
-				}
-				else if (dateTime.Month < anotherDateTime.Month)
-				{
-					return -1;
-				}
-			}
-			if (compareFieldMin >= DateTimeField.Day)
-			{
-				if (dateTime.Day > anotherDateTime.Day)
-				{
-					return 1;
-				}
-				else if (dateTime.Day < anotherDateTime.Day)
-				{
-					return -1;
-				}
-			}
-			if (compareFieldMin >= DateTimeField.Hour)
-			{
-				if (dateTime.Hour > anotherDateTime.Hour)
-				{
-					return 1;
-				}
-				else if (dateTime.Hour < anotherDateTime.Hour)
-				{
-					return -1;
-				}
-			}
-			if (compareFieldMin >= DateTimeField.Minute)
-			{
-				if (dateTime.Minute > anotherDateTime.Minute)
-				{
-					return 1;
-				}
-				else if (dateTime.Minute < anotherDateTime.Minute)
-				{
-					return -1;
-				}
-			}
-			if (compareFieldMin >= DateTimeField.Second)
-			{
-				if (dateTime.Second > anotherDateTime.Second)
-				{
-					return 1;
-				}
-				else if (dateTime.Second < anotherDateTime.Second)
-				{
-					return -1;
-				}
-			}
-			if (compareFieldMin >= DateTimeField.Millisecond)
-			{
-				if (dateTime.Millisecond > anotherDateTime.Millisecond)
-				{
-					return 1;
-				}
-				else if (dateTime.Millisecond < anotherDateTime.Millisecond)
-				{
-					return -1;
-				}
-			}
-			return 0;
-		}
-
-
-		/// <summary>
-		/// 在指定的时间周期内比较两个时间对象，如：指定的时间范围为“Year”时，则比较时忽略时间对象的年份值。
-		/// </summary>
-		/// <param name="dateTime"></param>
-		/// <param name="anotherDateTime"></param>
-		/// <param name="compareCycle"></param>
-		/// <returns></returns>
-		public static int CompareTo(
-			this DateTime dateTime,
-			DateTime anotherDateTime,
-			DateTimeCycle compareCycle)
-		{
-			switch (compareCycle)
-			{
-				default:
-				case DateTimeCycle.All:
-				case DateTimeCycle.Century:
+					if (dateTime.Month < anotherDateTime.Month)
 					{
-						return dateTime.CompareTo(anotherDateTime);
+						return -1;
 					}
-				case DateTimeCycle.Year:
+					else if (dateTime.Month > anotherDateTime.Month)
 					{
-						if (dateTime.Month < anotherDateTime.Month)
-						{
-							return -1;
-						}
-						else if (dateTime.Month > anotherDateTime.Month)
-						{
-							return 1;
-						}
-						else if (dateTime.Day < anotherDateTime.Day)
-						{
-							return -1;
-						}
-						else if (dateTime.Day > anotherDateTime.Day)
-						{
-							return 1;
-						}
-						else if (dateTime.Hour < anotherDateTime.Hour)
-						{
-							return -1;
-						}
-						else if (dateTime.Hour > anotherDateTime.Hour)
-						{
-							return 1;
-						}
-						else if (dateTime.Minute < anotherDateTime.Minute)
-						{
-							return -1;
-						}
-						else if (dateTime.Minute > anotherDateTime.Minute)
-						{
-							return 1;
-						}
-						else if (dateTime.Second < anotherDateTime.Second)
-						{
-							return -1;
-						}
-						else if (dateTime.Second > anotherDateTime.Second)
-						{
-							return 1;
-						}
-						else if (dateTime.Millisecond < anotherDateTime.Millisecond)
-						{
-							return -1;
-						}
-						else if (dateTime.Millisecond > anotherDateTime.Millisecond)
-						{
-							return 1;
-						}
+						return 1;
 					}
-					break;
-				case DateTimeCycle.Month:
+					else if (dateTime.Day < anotherDateTime.Day)
 					{
-						if (dateTime.Day < anotherDateTime.Day)
-						{
-							return -1;
-						}
-						else if (dateTime.Day > anotherDateTime.Day)
-						{
-							return 1;
-						}
-						else if (dateTime.Hour < anotherDateTime.Hour)
-						{
-							return -1;
-						}
-						else if (dateTime.Hour > anotherDateTime.Hour)
-						{
-							return 1;
-						}
-						else if (dateTime.Minute < anotherDateTime.Minute)
-						{
-							return -1;
-						}
-						else if (dateTime.Minute > anotherDateTime.Minute)
-						{
-							return 1;
-						}
-						else if (dateTime.Second < anotherDateTime.Second)
-						{
-							return -1;
-						}
-						else if (dateTime.Second > anotherDateTime.Second)
-						{
-							return 1;
-						}
-						else if (dateTime.Millisecond < anotherDateTime.Millisecond)
-						{
-							return -1;
-						}
-						else if (dateTime.Millisecond > anotherDateTime.Millisecond)
-						{
-							return 1;
-						}
+						return -1;
 					}
-					break;
-				case DateTimeCycle.Week:
+					else if (dateTime.Day > anotherDateTime.Day)
 					{
-						if (dateTime.DayOfWeek < anotherDateTime.DayOfWeek)
-						{
-							return -1;
-						}
-						else if (dateTime.DayOfWeek > anotherDateTime.DayOfWeek)
-						{
-							return 1;
-						}
-						else if (dateTime.Hour < anotherDateTime.Hour)
-						{
-							return -1;
-						}
-						else if (dateTime.Hour > anotherDateTime.Hour)
-						{
-							return 1;
-						}
-						else if (dateTime.Minute < anotherDateTime.Minute)
-						{
-							return -1;
-						}
-						else if (dateTime.Minute > anotherDateTime.Minute)
-						{
-							return 1;
-						}
-						else if (dateTime.Second < anotherDateTime.Second)
-						{
-							return -1;
-						}
-						else if (dateTime.Second > anotherDateTime.Second)
-						{
-							return 1;
-						}
-						else if (dateTime.Millisecond < anotherDateTime.Millisecond)
-						{
-							return -1;
-						}
-						else if (dateTime.Millisecond > anotherDateTime.Millisecond)
-						{
-							return 1;
-						}
+						return 1;
 					}
-					break;
-				case DateTimeCycle.Day:
+					else if (dateTime.Hour < anotherDateTime.Hour)
 					{
-						if (dateTime.Hour < anotherDateTime.Hour)
-						{
-							return -1;
-						}
-						else if (dateTime.Hour > anotherDateTime.Hour)
-						{
-							return 1;
-						}
-						else if (dateTime.Minute < anotherDateTime.Minute)
-						{
-							return -1;
-						}
-						else if (dateTime.Minute > anotherDateTime.Minute)
-						{
-							return 1;
-						}
-						else if (dateTime.Second < anotherDateTime.Second)
-						{
-							return -1;
-						}
-						else if (dateTime.Second > anotherDateTime.Second)
-						{
-							return 1;
-						}
-						else if (dateTime.Millisecond < anotherDateTime.Millisecond)
-						{
-							return -1;
-						}
-						else if (dateTime.Millisecond > anotherDateTime.Millisecond)
-						{
-							return 1;
-						}
+						return -1;
 					}
-					break;
-				case DateTimeCycle.Hour:
+					else if (dateTime.Hour > anotherDateTime.Hour)
 					{
-						if (dateTime.Minute < anotherDateTime.Minute)
-						{
-							return -1;
-						}
-						else if (dateTime.Minute > anotherDateTime.Minute)
-						{
-							return 1;
-						}
-						else if (dateTime.Second < anotherDateTime.Second)
-						{
-							return -1;
-						}
-						else if (dateTime.Second > anotherDateTime.Second)
-						{
-							return 1;
-						}
-						else if (dateTime.Millisecond < anotherDateTime.Millisecond)
-						{
-							return -1;
-						}
-						else if (dateTime.Millisecond > anotherDateTime.Millisecond)
-						{
-							return 1;
-						}
+						return 1;
 					}
-					break;
-				case DateTimeCycle.Minute:
+					else if (dateTime.Minute < anotherDateTime.Minute)
 					{
-						if (dateTime.Second < anotherDateTime.Second)
-						{
-							return -1;
-						}
-						else if (dateTime.Second > anotherDateTime.Second)
-						{
-							return 1;
-						}
-						else if (dateTime.Millisecond < anotherDateTime.Millisecond)
-						{
-							return -1;
-						}
-						else if (dateTime.Millisecond > anotherDateTime.Millisecond)
-						{
-							return 1;
-						}
+						return -1;
 					}
-					break;
-				case DateTimeCycle.Second:
+					else if (dateTime.Minute > anotherDateTime.Minute)
 					{
-						if (dateTime.Millisecond < anotherDateTime.Millisecond)
-						{
-							return -1;
-						}
-						else if (dateTime.Millisecond > anotherDateTime.Millisecond)
-						{
-							return 1;
-						}
+						return 1;
 					}
-					break;
-				case DateTimeCycle.Millisecond:
+					else if (dateTime.Second < anotherDateTime.Second)
 					{
-						// !!!⚠ 毫秒以下不进行比较，永远相等。 ⚠!!!
+						return -1;
 					}
-					break;
-			}
-			return 0;
+					else if (dateTime.Second > anotherDateTime.Second)
+					{
+						return 1;
+					}
+					else if (dateTime.Millisecond < anotherDateTime.Millisecond)
+					{
+						return -1;
+					}
+					else if (dateTime.Millisecond > anotherDateTime.Millisecond)
+					{
+						return 1;
+					}
+				}
+				break;
+			case DateTimeCycle.Month:
+				{
+					if (dateTime.Day < anotherDateTime.Day)
+					{
+						return -1;
+					}
+					else if (dateTime.Day > anotherDateTime.Day)
+					{
+						return 1;
+					}
+					else if (dateTime.Hour < anotherDateTime.Hour)
+					{
+						return -1;
+					}
+					else if (dateTime.Hour > anotherDateTime.Hour)
+					{
+						return 1;
+					}
+					else if (dateTime.Minute < anotherDateTime.Minute)
+					{
+						return -1;
+					}
+					else if (dateTime.Minute > anotherDateTime.Minute)
+					{
+						return 1;
+					}
+					else if (dateTime.Second < anotherDateTime.Second)
+					{
+						return -1;
+					}
+					else if (dateTime.Second > anotherDateTime.Second)
+					{
+						return 1;
+					}
+					else if (dateTime.Millisecond < anotherDateTime.Millisecond)
+					{
+						return -1;
+					}
+					else if (dateTime.Millisecond > anotherDateTime.Millisecond)
+					{
+						return 1;
+					}
+				}
+				break;
+			case DateTimeCycle.Week:
+				{
+					if (dateTime.DayOfWeek < anotherDateTime.DayOfWeek)
+					{
+						return -1;
+					}
+					else if (dateTime.DayOfWeek > anotherDateTime.DayOfWeek)
+					{
+						return 1;
+					}
+					else if (dateTime.Hour < anotherDateTime.Hour)
+					{
+						return -1;
+					}
+					else if (dateTime.Hour > anotherDateTime.Hour)
+					{
+						return 1;
+					}
+					else if (dateTime.Minute < anotherDateTime.Minute)
+					{
+						return -1;
+					}
+					else if (dateTime.Minute > anotherDateTime.Minute)
+					{
+						return 1;
+					}
+					else if (dateTime.Second < anotherDateTime.Second)
+					{
+						return -1;
+					}
+					else if (dateTime.Second > anotherDateTime.Second)
+					{
+						return 1;
+					}
+					else if (dateTime.Millisecond < anotherDateTime.Millisecond)
+					{
+						return -1;
+					}
+					else if (dateTime.Millisecond > anotherDateTime.Millisecond)
+					{
+						return 1;
+					}
+				}
+				break;
+			case DateTimeCycle.Day:
+				{
+					if (dateTime.Hour < anotherDateTime.Hour)
+					{
+						return -1;
+					}
+					else if (dateTime.Hour > anotherDateTime.Hour)
+					{
+						return 1;
+					}
+					else if (dateTime.Minute < anotherDateTime.Minute)
+					{
+						return -1;
+					}
+					else if (dateTime.Minute > anotherDateTime.Minute)
+					{
+						return 1;
+					}
+					else if (dateTime.Second < anotherDateTime.Second)
+					{
+						return -1;
+					}
+					else if (dateTime.Second > anotherDateTime.Second)
+					{
+						return 1;
+					}
+					else if (dateTime.Millisecond < anotherDateTime.Millisecond)
+					{
+						return -1;
+					}
+					else if (dateTime.Millisecond > anotherDateTime.Millisecond)
+					{
+						return 1;
+					}
+				}
+				break;
+			case DateTimeCycle.Hour:
+				{
+					if (dateTime.Minute < anotherDateTime.Minute)
+					{
+						return -1;
+					}
+					else if (dateTime.Minute > anotherDateTime.Minute)
+					{
+						return 1;
+					}
+					else if (dateTime.Second < anotherDateTime.Second)
+					{
+						return -1;
+					}
+					else if (dateTime.Second > anotherDateTime.Second)
+					{
+						return 1;
+					}
+					else if (dateTime.Millisecond < anotherDateTime.Millisecond)
+					{
+						return -1;
+					}
+					else if (dateTime.Millisecond > anotherDateTime.Millisecond)
+					{
+						return 1;
+					}
+				}
+				break;
+			case DateTimeCycle.Minute:
+				{
+					if (dateTime.Second < anotherDateTime.Second)
+					{
+						return -1;
+					}
+					else if (dateTime.Second > anotherDateTime.Second)
+					{
+						return 1;
+					}
+					else if (dateTime.Millisecond < anotherDateTime.Millisecond)
+					{
+						return -1;
+					}
+					else if (dateTime.Millisecond > anotherDateTime.Millisecond)
+					{
+						return 1;
+					}
+				}
+				break;
+			case DateTimeCycle.Second:
+				{
+					if (dateTime.Millisecond < anotherDateTime.Millisecond)
+					{
+						return -1;
+					}
+					else if (dateTime.Millisecond > anotherDateTime.Millisecond)
+					{
+						return 1;
+					}
+				}
+				break;
+			case DateTimeCycle.Millisecond:
+				{
+					// !!!⚠ 毫秒以下不进行比较，永远相等。 ⚠!!!
+				}
+				break;
 		}
+		return 0;
+	}
 
-		public static bool IsEarlierInYear(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
+	public static bool IsEarlierInYear(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Year)
+			< 0;
+	}
+
+	public static bool EqualsInYear(
+		this DateTime dateTime,
+		DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Year)
+			== 0;
+	}
+
+	public static bool IsLaterInYear(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Year)
+			> 0;
+	}
+
+	public static bool IsEarlierInMonth(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Month)
+			< 0;
+	}
+
+	public static bool EqualsInMonth(
+		this DateTime dateTime,
+		DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Month)
+			== 0;
+	}
+
+	public static bool IsLaterInMonth(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Month)
+			> 0;
+	}
+
+	public static bool IsLastMonthOf(
+		this DateTime dateTime,
+	       DateTime currentMonth)
+	{
+		var dateTimeBeforeMonth = currentMonth.AddMonths(-1);
+		if (dateTimeBeforeMonth.EqualsInMonth(dateTime))
 		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Year)
-				< 0;
+			return true;
 		}
+		return false;
+	}
 
-		public static bool EqualsInYear(
-			this DateTime dateTime,
-			DateTime anotherDateTime)
+	public static bool IsNextMonthOf(
+		this DateTime dateTime,
+	       DateTime currentMonth)
+	{
+		var dateTimeAfterMonth = currentMonth.AddMonths(1);
+		if (dateTimeAfterMonth.EqualsInMonth(dateTime))
 		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Year)
-				== 0;
+			return true;
 		}
+		return false;
+	}
 
-		public static bool IsLaterInYear(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
+	public static bool IsLastWeekInStartsWithMondayOf(
+		this DateTime dateTime,
+	       DateTime currentWeek)
+	{
+		var currentWeekBeginTime
+			= currentWeek.FirstDayOfThisWeekStartsWithMonday();
+		var lastWeekBeginTime
+			= currentWeek.AddDays(-1 * TimeConstants.DaysToWeek);
+		if (dateTime >= lastWeekBeginTime
+			&& dateTime < currentWeekBeginTime)
 		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Year)
-				> 0;
+			return true;
 		}
+		return false;
+	}
 
-		public static bool IsEarlierInMonth(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
+	public static bool IsNextWeekInStartsWithMondayOf(
+		this DateTime dateTime,
+	       DateTime currentWeek)
+	{
+		var nextWeekBeginTime
+			= currentWeek.FirstDayOfNextWeekStartsWithMonday();
+		var nextWeekEndTime
+			= nextWeekBeginTime.AddDays(TimeConstants.DaysToWeek);
+		if (dateTime >= nextWeekBeginTime
+			&& dateTime < nextWeekEndTime)
 		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Month)
-				< 0;
+			return true;
 		}
+		return false;
+	}
 
-		public static bool EqualsInMonth(
-			this DateTime dateTime,
-			DateTime anotherDateTime)
+	public static bool IsEarlierInDay(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Day)
+			< 0;
+	}
+
+	public static bool EqualsInDay(
+		this DateTime dateTime,
+		DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Day)
+			== 0;
+	}
+
+	public static bool IsLaterInDay(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Day)
+			> 0;
+	}
+
+	public static bool IsLastDayOf(this DateTime dateTime, DateTime today)
+	{
+		var dateTimeBeforeDay = today.AddDays(-1);
+		if (dateTimeBeforeDay.EqualsInDay(dateTime))
 		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Month)
-				== 0;
+			return true;
 		}
-
-		public static bool IsLaterInMonth(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
+		return false;
+	}
+	public static bool IsNextDayOf(this DateTime dateTime, DateTime today)
+	{
+		var dateTimeAfterDay = today.AddDays(1);
+		if (dateTimeAfterDay.EqualsInDay(dateTime))
 		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Month)
-				> 0;
+			return true;
 		}
+		return false;
+	}
 
-		public static bool IsLastMonthOf(
-			this DateTime dateTime,
-		       DateTime currentMonth)
+	public static bool IsYesterdayOf(
+		this DateTime dateTime,
+	       DateTime today)
+	{
+		return dateTime.IsLastDayOf(today);
+	}
+
+	public static bool IsTomorrowOf(
+		this DateTime dateTime,
+	       DateTime today)
+	{
+		return dateTime.IsNextDayOf(today);
+	}
+
+	public static bool IsEarlierInHour(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Hour)
+			< 0;
+	}
+
+	public static bool EqualsInHour(
+		this DateTime dateTime,
+		DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Hour)
+			== 0;
+	}
+
+	public static bool IsLaterInHour(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Hour)
+			> 0;
+	}
+
+	public static bool IsLastHourOf(this DateTime dateTime, DateTime currentDateTime)
+	{
+		if (dateTime.AddHours(1).EqualsInHour(currentDateTime))
 		{
-			var dateTimeBeforeMonth = currentMonth.AddMonths(-1);
-			if (dateTimeBeforeMonth.EqualsInMonth(dateTime))
+			return true;
+		}
+		return false;
+	}
+
+	public static bool IsNextHourOf(this DateTime dateTime, DateTime currentDateTime)
+	{
+		if (dateTime.AddHours(-1).EqualsInHour(currentDateTime))
+		{
+			return true;
+		}
+		return false;
+	}
+
+
+	public static bool IsEarlierInMinute(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Minute)
+			< 0;
+	}
+
+	public static bool EqualsInMinute(
+		this DateTime dateTime,
+		DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Minute)
+			== 0;
+	}
+
+	public static bool IsLaterInMinute(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Minute)
+			> 0;
+	}
+
+	public static bool IsLastMinuteOf(this DateTime dateTime, DateTime currentDateTime)
+	{
+		if (dateTime.AddMinutes(1).EqualsInMinute(currentDateTime))
+		{
+			return true;
+		}
+		return false;
+	}
+
+	public static bool IsNextMinuteOf(this DateTime dateTime, DateTime currentDateTime)
+	{
+		if (dateTime.AddMinutes(-1).EqualsInMinute(currentDateTime))
+		{
+			return true;
+		}
+		return false;
+	}
+
+	public static bool IsEarlierInSecond(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Second)
+			< 0;
+	}
+
+	public static bool EqualsInSecond(
+		this DateTime dateTime,
+		DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Second)
+			== 0;
+	}
+
+	public static bool IsLaterInSecond(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Second)
+			> 0;
+	}
+
+	public static bool IsLastSecondOf(this DateTime dateTime, DateTime currentDateTime)
+	{
+		if (dateTime.AddSeconds(1).EqualsInSecond(currentDateTime))
+		{
+			return true;
+		}
+		return false;
+	}
+
+	public static bool IsNextSecondOf(this DateTime dateTime, DateTime currentDateTime)
+	{
+		if (dateTime.AddSeconds(-1).EqualsInSecond(currentDateTime))
+		{
+			return true;
+		}
+		return false;
+	}
+
+	public static bool IsEarlierInMillisecond(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Millisecond)
+			< 0;
+	}
+
+	public static bool EqualsInMillisecond(
+		this DateTime dateTime,
+		DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Millisecond)
+			== 0;
+	}
+
+	public static bool IsLaterInMillisecond(
+	       this DateTime dateTime,
+	       DateTime anotherDateTime)
+	{
+		return dateTime.CompareTo(
+			anotherDateTime,
+			DateTimeField.Millisecond)
+			> 0;
+	}
+	public static bool IsLastMillisecondOf(this DateTime dateTime, DateTime currentDateTime)
+	{
+		if (dateTime.AddMilliseconds(1).EqualsInMillisecond(currentDateTime))
+		{
+			return true;
+		}
+		return false;
+	}
+
+	public static bool IsNextMillisecondOf(this DateTime dateTime, DateTime currentDateTime)
+	{
+		if (dateTime.AddMilliseconds(-1).EqualsInMillisecond(currentDateTime))
+		{
+			return true;
+		}
+		return false;
+	}
+
+	public static bool IsBetween(
+		this DateTime dateTime,
+		DateTime? beginTime,
+		DateTime? endTime,
+		DateTimeField compareFieldMin)
+	{
+		if (beginTime != null
+			&& endTime != null)
+		{
+			if (dateTime.CompareTo(beginTime.Value, compareFieldMin) >= 0
+				&& dateTime.CompareTo(endTime.Value, compareFieldMin) < 0)
 			{
 				return true;
 			}
-			return false;
 		}
-
-		public static bool IsNextMonthOf(
-			this DateTime dateTime,
-		       DateTime currentMonth)
+		else if (beginTime != null)
 		{
-			var dateTimeAfterMonth = currentMonth.AddMonths(1);
-			if (dateTimeAfterMonth.EqualsInMonth(dateTime))
+			if (dateTime.CompareTo(beginTime.Value, compareFieldMin) >= 0)
 			{
 				return true;
 			}
-			return false;
 		}
-
-		public static bool IsLastWeekInStartsWithMondayOf(
-			this DateTime dateTime,
-		       DateTime currentWeek)
+		else if (endTime != null)
 		{
-			var currentWeekBeginTime
-				= currentWeek.FirstDayOfThisWeekStartsWithMonday();
-			var lastWeekBeginTime
-				= currentWeek.AddDays(-1 * TimeConstants.DaysToWeek);
-			if (dateTime >= lastWeekBeginTime
-				&& dateTime < currentWeekBeginTime)
+			if (dateTime.CompareTo(endTime.Value, compareFieldMin) < 0)
 			{
 				return true;
 			}
-			return false;
 		}
+		return false;
+	}
 
-		public static bool IsNextWeekInStartsWithMondayOf(
-			this DateTime dateTime,
-		       DateTime currentWeek)
+	public static bool IsContinuousAfter(
+		this DateTime currentDateTime,
+		DateTime lastDateTime,
+		DateTimeField continuousAccuracy,
+		bool isEqualsToContinuous = true)
+	{
+		switch (continuousAccuracy)
 		{
-			var nextWeekBeginTime
-				= currentWeek.FirstDayOfNextWeekStartsWithMonday();
-			var nextWeekEndTime
-				= nextWeekBeginTime.AddDays(TimeConstants.DaysToWeek);
-			if (dateTime >= nextWeekBeginTime
-				&& dateTime < nextWeekEndTime)
-			{
-				return true;
-			}
-			return false;
-		}
-
-		public static bool IsEarlierInDay(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Day)
-				< 0;
-		}
-
-		public static bool EqualsInDay(
-			this DateTime dateTime,
-			DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Day)
-				== 0;
-		}
-
-		public static bool IsLaterInDay(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Day)
-				> 0;
-		}
-		public static bool IsYesterdayOf(
-			this DateTime dateTime,
-		       DateTime today)
-		{
-			var dateTimeBeforeDay = today.AddDays(-1);
-			if (dateTimeBeforeDay.EqualsInDay(dateTime))
-			{
-				return true;
-			}
-			return false;
-		}
-
-		public static bool IsTomorrowOf(
-			this DateTime dateTime,
-		       DateTime today)
-		{
-			var dateTimeAfterDay = today.AddDays(1);
-			if (dateTimeAfterDay.EqualsInDay(dateTime))
-			{
-				return true;
-			}
-			return false;
-		}
-
-		public static bool IsEarlierInHour(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Hour)
-				< 0;
-		}
-
-		public static bool EqualsInHour(
-			this DateTime dateTime,
-			DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Hour)
-				== 0;
-		}
-
-		public static bool IsLaterInHour(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Hour)
-				> 0;
-		}
-
-
-		public static bool IsEarlierInMinute(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Minute)
-				< 0;
-		}
-
-		public static bool EqualsInMinute(
-			this DateTime dateTime,
-			DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Minute)
-				== 0;
-		}
-
-		public static bool IsLaterInMinute(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Minute)
-				> 0;
-		}
-
-
-		public static bool IsEarlierInSecond(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Second)
-				< 0;
-		}
-
-		public static bool EqualsInSecond(
-			this DateTime dateTime,
-			DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Second)
-				== 0;
-		}
-
-		public static bool IsLaterInSecond(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Second)
-				> 0;
-		}
-
-
-		public static bool IsEarlierInMillisecond(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Millisecond)
-				< 0;
-		}
-
-		public static bool EqualsInMillisecond(
-			this DateTime dateTime,
-			DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Millisecond)
-				== 0;
-		}
-
-		public static bool IsLaterInMillisecond(
-		       this DateTime dateTime,
-		       DateTime anotherDateTime)
-		{
-			return dateTime.CompareTo(
-				anotherDateTime,
-				DateTimeField.Millisecond)
-				> 0;
-		}
-
-		public static bool IsBetween(
-			this DateTime dateTime,
-			DateTime? beginTime,
-			DateTime? endTime,
-			DateTimeField compareFieldMin)
-		{
-			if (beginTime != null
-				&& endTime != null)
-			{
-				if (dateTime.CompareTo(beginTime.Value, compareFieldMin) >= 0
-					&& dateTime.CompareTo(endTime.Value, compareFieldMin) < 0)
+			case DateTimeField.Year:
 				{
-					return true;
+					if (currentDateTime.Year == lastDateTime.Year + 1)
+					{
+						return true;
+					}
+					else if (currentDateTime.Year == lastDateTime.Year)
+					{
+						if (isEqualsToContinuous)
+						{
+							return true;
+						}
+					}
 				}
-			}
-			else if (beginTime != null)
-			{
-				if (dateTime.CompareTo(beginTime.Value, compareFieldMin) >= 0)
+				break;
+			case DateTimeField.Month:
 				{
-					return true;
+					if (currentDateTime.IsNextMonthOf(lastDateTime))
+					{
+						return true;
+					}
+					else if (currentDateTime.EqualsInMonth(lastDateTime))
+					{
+						if (isEqualsToContinuous)
+						{
+							return true;
+						}
+					}
 				}
-			}
-			else if (endTime != null)
-			{
-				if (dateTime.CompareTo(endTime.Value, compareFieldMin) < 0)
+				break;
+			case DateTimeField.Day:
 				{
-					return true;
+					if (currentDateTime.IsNextDayOf(lastDateTime))
+					{
+						return true;
+					}
+					else if (currentDateTime.EqualsInDay(lastDateTime))
+					{
+						if (isEqualsToContinuous)
+						{
+							return true;
+						}
+					}
 				}
-			}
-			return false;
+				break;
+			case DateTimeField.Hour:
+				{
+					if (currentDateTime.IsNextHourOf(lastDateTime))
+					{
+						return true;
+					}
+					else if (currentDateTime.EqualsInHour(lastDateTime))
+					{
+						if (isEqualsToContinuous)
+						{
+							return true;
+						}
+					}
+				}
+				break;
+			case DateTimeField.Minute:
+				{
+					if (currentDateTime.IsNextMinuteOf(lastDateTime))
+					{
+						return true;
+					}
+					else if (currentDateTime.EqualsInMinute(lastDateTime))
+					{
+						if (isEqualsToContinuous)
+						{
+							return true;
+						}
+					}
+				}
+				break;
+			case DateTimeField.Second:
+				{
+					if (currentDateTime.IsNextSecondOf(lastDateTime))
+					{
+						return true;
+					}
+					else if (currentDateTime.EqualsInSecond(lastDateTime))
+					{
+						if (isEqualsToContinuous)
+						{
+							return true;
+						}
+					}
+				}
+				break;
+			case DateTimeField.Millisecond:
+				{
+					if (currentDateTime.IsNextMillisecondOf(lastDateTime))
+					{
+
+						return true;
+					}
+					else if (currentDateTime.EqualsInMillisecond(lastDateTime))
+					{
+						if (isEqualsToContinuous)
+						{
+							return true;
+						}
+					}
+				}
+				break;
+			default:
+				{ }
+				break;
 		}
+		return false;
+	}
 
 
-		public static string ToListElementCaptionDefault(
-			this DateTime dateTime,
-			bool isNeedSecondsField = false)
+	public static bool IsContinuousBefore(
+		this DateTime currentDateTime,
+		DateTime nextDateTime,
+		DateTimeField continuousAccuracy,
+		bool isEqualsToContinuous = true)
+	{
+		return nextDateTime.IsContinuousAfter(
+			currentDateTime,
+			continuousAccuracy,
+			isEqualsToContinuous);
+	}
+
+	public static string CaptionOfListElementDefault(
+		this DateTime dateTime,
+		bool isNeedSecondsField = false)
+	{
+		string caption;
+		var now = DateTime.Now;
+		if (dateTime.EqualsInDay(now))
 		{
-			string caption;
-			var now = DateTime.Now;
-			if (dateTime.EqualsInDay(now))
+			if (isNeedSecondsField)
 			{
-				if (isNeedSecondsField)
-				{
-					caption = dateTime.ToString("今日 HH:mm:ss");
-				}
-				else
-				{
-					caption = dateTime.ToString("今日 HH:mm");
-				}
+				caption = dateTime.ToString("今天 HH:mm:ss");
 			}
-			else if (dateTime.IsYesterdayOf(now))
+			else
 			{
-				if (isNeedSecondsField)
-				{
-					caption = dateTime.ToString("昨日 HH:mm:ss");
-				}
-				else
-				{
-					caption = dateTime.ToString("昨日 HH:mm");
-				}
+				caption = dateTime.ToString("今天 HH:mm");
 			}
-			else if (dateTime.IsTomorrowOf(now))
+		}
+		else if (dateTime.IsYesterdayOf(now))
+		{
+			if (isNeedSecondsField)
 			{
-				if (isNeedSecondsField)
-				{
-					caption = dateTime.ToString("明日 HH:mm:ss");
-				}
-				else
-				{
-					caption = dateTime.ToString("明日 HH:mm");
-				}
+				caption = dateTime.ToString("昨天 HH:mm:ss");
 			}
-			else if (dateTime.EqualsInYear(now))
+			else
+			{
+				caption = dateTime.ToString("昨天 HH:mm");
+			}
+		}
+		else if (dateTime.IsTomorrowOf(now))
+		{
+			if (isNeedSecondsField)
+			{
+				caption = dateTime.ToString("明天 HH:mm:ss");
+			}
+			else
+			{
+				caption = dateTime.ToString("明天 HH:mm");
+			}
+		}
+		else if (dateTime.EqualsInYear(now))
+		{
+			caption = dateTime.ToString("MM月dd天 HH:mm");
+		}
+		else
+		{
+			caption = dateTime.ToString("yyyy年MM月dd日");
+		}
+		return caption;
+	}
+
+	public static string CaptionOfDetailPageDefault(
+		this DateTime dateTime,
+		bool isNeedSecondsField = false)
+	{
+		string caption;
+		var now = DateTime.Now;
+		if (dateTime.EqualsInDay(now))
+		{
+			if (isNeedSecondsField)
+			{
+				caption = dateTime.ToString("今天 HH:mm:ss");
+			}
+			else
+			{
+				caption = dateTime.ToString("今天 HH:mm");
+			}
+		}
+		else if (dateTime.IsYesterdayOf(now))
+		{
+			if (isNeedSecondsField)
+			{
+				caption = dateTime.ToString("昨天 HH:mm:ss");
+			}
+			else
+			{
+				caption = dateTime.ToString("昨天 HH:mm");
+			}
+		}
+		else if (dateTime.IsTomorrowOf(now))
+		{
+			if (isNeedSecondsField)
+			{
+				caption = dateTime.ToString("明天 HH:mm:ss");
+			}
+			else
+			{
+				caption = dateTime.ToString("明天 HH:mm");
+			}
+		}
+		else if (dateTime.EqualsInYear(now))
+		{
+			if (isNeedSecondsField)
+			{
+				caption = dateTime.ToString("MM月dd日 HH:mm:ss");
+			}
+			else
 			{
 				caption = dateTime.ToString("MM月dd日 HH:mm");
+			}
+		}
+		else
+		{
+			if (isNeedSecondsField)
+			{
+				caption = dateTime.ToString("yyyy年MM月dd日 HH:mm:ss");
 			}
 			else
 			{
 				caption = dateTime.ToString("yyyy年MM月dd日");
 			}
-			return caption;
 		}
-
-		public static string ToDetailPageCaptionDefault(
-			this DateTime dateTime,
-			bool isNeedSecondsField = false)
-		{
-			string caption;
-			var now = DateTime.Now;
-			if (dateTime.EqualsInDay(now))
-			{
-				if (isNeedSecondsField)
-				{
-					caption = dateTime.ToString("今日 HH:mm:ss");
-				}
-				else
-				{
-					caption = dateTime.ToString("今日 HH:mm");
-				}
-			}
-			else if (dateTime.IsYesterdayOf(now))
-			{
-				if (isNeedSecondsField)
-				{
-					caption = dateTime.ToString("昨日 HH:mm:ss");
-				}
-				else
-				{
-					caption = dateTime.ToString("昨日 HH:mm");
-				}
-			}
-			else if (dateTime.IsTomorrowOf(now))
-			{
-				if (isNeedSecondsField)
-				{
-					caption = dateTime.ToString("明日 HH:mm:ss");
-				}
-				else
-				{
-					caption = dateTime.ToString("明日 HH:mm");
-				}
-			}
-			else if (dateTime.EqualsInYear(now))
-			{
-				if (isNeedSecondsField)
-				{
-					caption = dateTime.ToString("MM月dd日 HH:mm:ss");
-				}
-				else
-				{
-					caption = dateTime.ToString("MM月dd日 HH:mm");
-				}
-			}
-			else
-			{
-				if (isNeedSecondsField)
-				{
-					caption = dateTime.ToString("yyyy年MM月dd日 HH:mm:ss");
-				}
-				else
-				{
-					caption = dateTime.ToString("yyyy年MM月dd日");
-				}
-			}
-			return caption;
-		}
+		return caption;
 	}
+
 }
